Issue 2 add_markers not simulate node
I see that you have only one add_markers node that you use in home_service.sh
and add_markers.sh files. But the requirement of add_marker.sh file, is that
the marker should be published in the pickup location then you must delete the
marker after 5 sec and then republish the marker in the drop off . All this
must happen based on time and not based on any topic.

Solution: Add cube_node_simulate